requirements.txt
pandas
datasets
qdrant-client
sentence-transformers
langchain-text-splitters
pydantic-settings
python-dotenv

llama-index
llama-index-llms-gemini
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
ragas
arize-phoenix
pytest
pytest-asyncio
fastapi
uvicorn
sse-starlette
httpx
aiosqlite
